rename an remove safeareaview
